wayworn-hush: three findings the build cache had been hiding, and the tests directory gets the override every other game already has -- five of the seven were TEST_CASE bodies scored for cognitive complexity, which is measuring the wrong thing in a file whose job is to accumulate assertions, and wayworn was the one game of four missing the tests/.clang-tidy that says so; ReadingColor subtracted in int and let the divide widen it; and WorldClock's reported 24 * 60 was one of SIX places in that file pulling a day apart with bare numbers, so the day is read off one set of named constants now and the int-multiplication-then-widen goes away as a consequence rather than being cast quiet; none of this appeared on the earlier run because CI lints changed files PLUS every source beside a changed header, so touching one header pulled in all of wayworn's src and linted it for the first time

+++ games/wayworn-hush/include/ReadingColor.h
@@ -24,7 +24,7 @@ inline Color forReading(const growth::GrowthState& growth, const std::string& fa
 
     const growth::Rgb hue = growth::facultyColor(growth, faculty);
     // Difficulty 1..5 -> brightness ramp: common reads dim, legendary vivid.
-    const float t = std::clamp((difficulty - 1) / 4.0f, 0.0f, 1.0f);
+    const float t = std::clamp(static_cast<float>(difficulty - 1) / 4.0f, 0.0f, 1.0f);
     const float bright = 0.55f + 0.45f * t; // 0.55 (dim) .. 1.0 (vivid)
     return {hue.r * bright, hue.g * bright, hue.b * bright, alpha};
 }
